// client/src/types/interfaces.ts
export interface IReviewEndpoints {
  list: string;
  add: string;
  remove: (val: any) => string;
}

export interface IReviewadd {
  mediaId: number;
  mediaType: string;
  mediaTitle: string;
  mediaPoster: string;
  content: string;
}
export interface IReviewID {
  reviewId: number;
}